requirements.txt
Flask==2.3.2
openai==0.27.8
pymongo==4.5.0
python-dotenv==0.19.1